Single line print info of subvolume configuration
The line breaks are only printed when stdout_callback=debug is set in
ansible.cfg, which is not the default. For the default stdout_callback
line breaks are just printed as \n. In that case the information is not
easy to read. Therefore only a single line is printed now. Using it in a
single line also saves space in the logs.