Refactor: Set internal variables in subvolume.yaml
Instead of evaluating the template variables for a subvolume on a task
level, they are now all set in the beginning. This is cleaner and
easier to maintain. To distinguish from other role variables the
internal variables are prefixed with a subvolume_i_ (i for internal).

This commit also adds a new feature:

  - A subvolume definition can now contain an own device_uuid. When none
    is given the value from btrfssubvol_device_uuid is taken. This
    enables the management of multiple btrfs devices.