Umount unmanaged subvolumes
It is tried to unmount mounts of the subvolume, which are not managed by
this role.

Errors in unmounting are ignored, as the leftover mounts do no big harm
and most of the times it is "target is busy". The state=absent of the
mount task will still remove the entries from /etc/fstab. After a reboot
of the system that mount point will be gone.

Nevertheless, state=absent does not remove mounts, which have no entries
in /etc/fstab. Therefore a second task is added to remove such mounts.