Enhance UUID checks
The new check with ansible.builtin.match ensures correct UUID through
the regex. The regex supports lower and uppercase letters as defined in
RFC 4122.

* https://www.ietf.org/rfc/rfc4122.txt

Additionally it is now optional to set the btrfssubvol_device_uuid. But
in that case it is required to set the UUID on subvolume level.

This finishes the work of the previous commit, to prevent the role from
failing when included without configuration.