Skip mounting when mount point is empty string
When the mount_point is set to an empty string now the initial mount
management is skipped. The advantage of that is that now all mounts of
the subvolume are considered as unmanaged mounts. As a result the
tasks for unmounting the unmanaged mounts are triggered, which have
error handling. With this implementation it is not necessary to
implement duplicate error handling.

In this case it is also now necessary to create a mount point when the
mount_state is present. Setting permissions is also not possible, so
this task is skipped too.